Add support Visual Studio 2019 and Cygwin (gcc 9.3) Build configuration using Cmake (3.17.1)

+++ samples/ex-thread.c
@@ -17,11 +17,11 @@
 
 void *foo(void *data)
 {
-	int val = (unsigned long)data;
+	int val = (int)(uintptr_t)data;
 	printf( "Foo=%d\n", val );
 	//for ( ;; )
 		printf( "Hello %d\n", val );
-	
+
 	return 0;
 }
